Update collections-and-entries.md - Typos and Ordering
I fixed some typos, but I also think the "Ordering" section needs to be revisited. I believe the ordering of a collection has been moved into the "Structures" YAML file based on the Cool Writings starter kit and basic Statamic v3 installation. I was going to try to rewrite the section, but I am not sure what you want to do with the "Ordering" section on this page.