Update collections.md (#660)
* Update collections.md

Had someone in the discord misunderstanding the instructions for mounting - had a look myself and it did feel confusing. Made some minor tweaks to (hopefully) make it clearer?

* Update collections.md

Might be worth abstracting away to anything thats not blogs/articles/entries in order to explain the relationship without confusion? (not necessarily sure this is better).

* Explain both options on how to mount. Also, "fishies" is funny.

Co-authored-by: Jack McDade <jack@jackmcdade.com>